requirements.txt
aiodns==3.1.1
aiohttp==3.9.3
aiohttp-socks==0.8.4
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
ansible==7.5.0
ansible-core==2.14.5
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
backcall==0.2.0
bcolz-zipline==1.2.6
beautifulsoup4==4.11.2
bidict==0.23.0
bleach==6.0.0
blinker==1.6.2
boto3==1.28.14
botocore==1.31.14
cachetools==5.3.0
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cryptography==40.0.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
distro==1.9.0
dnspython==2.3.0
duckdb==0.7.1
elastic-transport==8.12.0
elasticsearch==8.12.0
email-validator==2.1.0.post1
entrypoints==0.4
exceptiongroup==1.1.3
executing==1.2.0
fake-useragent==1.4.0
fastjsonschema==2.16.3
fernet==1.0.1
filelock==3.7.1
Flask==2.1.2
Flask-Cors==4.0.0
Flask-Dance==7.0.0
Flask-Login==0.6.2
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
Fuzzy==1.2.2
geneimpacts==0.3.7
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.10
GitPython==3.1.31
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
googletransx==2.4.2
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
httpcore==1.0.3
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.3
idna==3.4
importlib-metadata==4.11.4
importlib-resources==5.12.0
inheritance==0.1.5
ipykernel==6.21.2
ipython==8.10.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.2
jupyter-events==0.6.3
jupyter_client==8.0.3
jupyter_core==5.2.0
jupyter_server==2.3.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.1
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
mock-open==1.4.0
multidict==6.0.5
mysql-connector==2.2.9
mysql-connector-python==8.0.33
nbclassic==0.5.2
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.22.4
oauthlib==3.2.2
openai==1.12.0
opencv-python==4.6.0.66
openpyxl==1.8.6
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
php==1.2.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.5.2
prometheus-client==0.16.0
prompt-toolkit==3.0.37
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyaes==1.6.1
pyarrow==11.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycares==4.4.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydot==1.4.2
pygame==2.1.3
Pygments==2.14.0
pylance==0.3.17
pymongo==4.1.1
Pympler==1.0.1
PyMySQL==1.0.3
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
PyQt5-sip==12.11.1
pyrsistent==0.19.3
PySocks==1.7.1
pyspark==3.4.0
python-dateutil==2.8.2
python-engineio==4.9.0
python-json-logger==2.0.7
python-socketio==5.11.1
python-socks==2.4.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
qrcode==7.4.2
qtconsole==5.4.0
QtPy==2.3.0
qtwidgets==1.1
regex==2023.12.25
requests==2.28.2
requests-oauthlib==1.3.1
resolvelib==0.8.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.5
rsa==4.9
s3transfer==0.6.1
safetensors==0.4.2
schedule==1.2.1
selenium==4.11.2
Send2Trash==1.8.0
simple-websocket==1.0.0
simplegeneric==0.8.1
six==1.16.0
sklearn==0.0.post4
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
stack-data==0.6.2
streamlit==1.22.0
tenacity==8.2.2
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.66.2
traitlets==5.9.0
transformers==4.37.2
trio==0.22.2
trio-websocket==0.10.3
twint==2.1.20
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==1.26.15
URLObject==2.4.3
validators==0.20.0
virtualenv==20.14.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.1.2
widgetsnbextension==4.0.5
wsproto==1.2.0
WTForms==3.1.0
yarl==1.9.4
zipp==3.8.0
